updates_BaseCalc_to_double

+++ src/at/sarah/oo/Calculator/BaseCalculator.java
@@ -2,19 +2,19 @@ package at.sarah.oo.Calculator;
 
 public class BaseCalculator {
 
-    public double add (int a, int b){
+    public double add (double a, double b){
         return a + b;
     }
 
-    public double subtract (int a, int b){
+    public double subtract (double a, double b){
         return a - b;
     }
 
-    public double divide (int a, int b){
+    public double divide (double a, double b){
         return (double) a / b;
     }
 
-    public double multiply (int a, int b){
+    public double multiply (double a, double b){
         return a * b;
     }
 }
